Fragments update @belal
