+++ .idea/misc.xml
@@ -3,12 +3,12 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-
+        <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/drawable/ic_baseline_star_border_24.xml" value="0.17395833333333333" />
+        <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/layout/activity_main.xml" value="0.16666666666666666" />
         <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/layout/fragment_dialog.xml" value="0.1578125" />
         <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/layout/fragment_favorite.xml" value="0.14375" />
-
-        <entry key="..\:/Users/MSI/Desktop/GeoGalleryProject/app/src/main/res/layout/activity_splash.xml" value="0.2828125" />
-
+        <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/layout/fragment_home_.xml" value="0.1578125" />
+        <entry key="..\:/Users/albaq/Desktop/Geo-Gallery-Project/app/src/main/res/menu/bottom_nav.xml" value="0.15462962962962962" />
       </map>
     </option>
   </component>
