+++ app/src/main/java/com/example/geogalleryproject/view/MainActivity.kt
@@ -1,12 +1,12 @@
-package com.example.geogalleryproject.view.main
+package com.example.geogalleryproject.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.geogalleryproject.R
 
-class Splash : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
+        setContentView(R.layout.activity_main)
     }
 }
