+++ app/src/main/java/com/example/geogalleryproject/view/Splash.kt
@@ -1,12 +1,12 @@
-package com.example.geogalleryproject.view.main
+package com.example.geogalleryproject.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.geogalleryproject.R
 
-class MainActivity : AppCompatActivity() {
+class Splash : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_splash)
     }
 }
