+++ app/src/main/java/com/example/geogalleryproject/view/main/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.geogalleryproject.view
+package com.example.geogalleryproject.view.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.geogalleryproject.R
 
-class FavoriteFragment : Fragment() {
+class HomeFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favorite, container, false)
+        return inflater.inflate(R.layout.fragment_home_, container, false)
     }
 
 }
