+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/HomeFragment"
+        android:icon="@drawable/ic_baseline_home_24"
+        android:title="@string/home" />
+    <item
+        android:id="@+id/FavoriteFragment"
+        android:icon="@drawable/ic_baseline_star_border_24"
+        android:title="@string/favorite" />
+</menu>
