+++ app/src/main/res/values/strings.xml
@@ -13,5 +13,7 @@
 
     <string name="hello_first_fragment">Hello first fragment</string>
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
+    <string name="favorite">Favorite</string>
+    <string name="home">Home</string>
 
 </resources>
